+++ R/broom.helpers-package.R
@@ -83,7 +83,7 @@ utils::globalVariables("where")
       dplyr::any_of(
         c(
           "y.level", "term", "variable", "var_label", "var_class", "var_type",
-          "header_row", "contrasts", "reference_row", "label"
+          "var_nlevels", "header_row", "contrasts", "reference_row", "label"
         )
       ),
       dplyr::everything()