+++ R/tidy_identify_variables.R
@@ -1,13 +1,17 @@
 #' Identify the variable corresponding to each model coefficient
 #'
 #' `tidy_identify_variables()` will add to the tidy tibble
-#' three additional columns: `variable`, `var_class` and `var_type`.
+#' three additional columns: `variable`, `var_class`, `var_type` and `var_nlevels`.
 #'
 #' It will also identify interaction terms and intercept(s).
+#'
 #' `var_type` could be `"continuous"`, `"categorical"`, `"intercept"`
 #' or `"interaction"`. Will be equal to `"unknown"` in the rare cases
 #' where `tidy_identify_variables()` will fail to identify the list
 #' of variables.
+#'
+#' For categorical variables, `var_nlevels` corresponds to the number
+#' of original levels in the corresponding variables.
 #' @param x a tidy tibble
 #' @param model the corresponding model, if not attached to `x`
 #' @inheritParams tidy_plus_plus
@@ -77,7 +81,8 @@ tidy_identify_variables <- function(x, model = tidy_get_model(x),
       dplyr::mutate(
         variable = NA_character_,
         var_class = NA_integer_,
-        var_type = "unknown"
+        var_type = "unknown",
+        var_nlevels = NA_integer_
       ) %>%
       tidy_attach_model(model = model, .attributes = .attributes)
   }